OF-2414: Where appropriate, send a stream error before closing a stream.
When a stream is being closed, it's often helpful to signal why this happens, especially when the stream is closed due to an error.

This commit adds stream errors in various places where such context is desirable.

The API to close a connection has been modified to be able to include an optional stream error. This replaces an older structure where a stream error is first sent manually, followed by a call to the 'close' API. With this change, only one attempt to send data is made instead of two (one for the stream error, and another one for the end-stream tag).